requirements.txt
langgraph 
langchain-openai 
python-dotenv
langchain-community
transformers
langchain_huggingface
streamlit